nixos-modules/microvm/ssh-deploy: fix for pristine deployments

+++ nixos-modules/microvm/ssh-deploy.nix
@@ -137,7 +137,9 @@ in
       if [[ $(realpath ./current) != $(realpath ./new) ]]; then
         echo "Installing MicroVM ${hostName}"
         rm -f old
-        mv current old
+        if [ -e current ]; then
+          mv current old
+        fi
         mv new current
 
         echo "Success. Diff:"
